backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Database - Using SQLite for development, can be changed to PostgreSQL
# psycopg2-binary==2.9.7  # Uncomment for PostgreSQL

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7
hashlib

# Environment & Configuration
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# Data Processing & Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
tensorflow==2.15.0
keras==2.15.0

# Web Scraping & Monitoring
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
fake-useragent==1.4.0

# Social Media APIs
python-telegram-bot==20.7
instagram-private-api==1.6.0
tweepy==4.14.0

# Real-time Communication
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# HTTP & API
urllib3==2.1.0
aiohttp==3.9.1
websockets==12.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.1.0
python-dotenv==1.0.0

# Monitoring & Logging
sentry-sdk==1.38.0
structlog==23.2.0
